% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_build_heatmap.R, R/sc_heatmap.R
\name{sc_build_heatmap}
\alias{sc_build_heatmap}
\title{Function for overlaying gene expression data onto dimensionality reduction results with the option of outlining cells based on known phenotype assignments}
\usage{
sc_build_heatmap(
  dataset = dataset,
  clusters = clusters,
  cluster.order = FALSE,
  gene.sets = gene.sets,
  col.space = 200,
  row.space = 1,
  rank = 25
)

sc_build_heatmap(
  dataset = dataset,
  clusters = clusters,
  cluster.order = FALSE,
  gene.sets = gene.sets,
  col.space = 200,
  row.space = 1,
  rank = 25
)
}
\arguments{
\item{dataset}{ExpressionSet object}

\item{dimred}{2-D output of t-SNE, UMAP, or PCA: Matrix with 2 columns corrersponding to dimensions 1 (x-axis) & 2 (y-axis), row numbers equivalent to the number of cells, and rownames corresponding to cell IDs}

\item{dimred.type}{Options include "tsne", "umap", or "pca"}

\item{gene.id.type}{Options include "ID" or "ENS" (ensembl ID)}

\item{pheno.var.sub}{Character; Cells are highlighted according to specified phenotype variable}

\item{pheno.var.sub.target}{Logical or Character; FALSE if default order; Vector of pheno.var sub-groups to be plotted}

\item{pheno.var.bg}{Logical or Character; FALSE if no additional subsetting, otherwise specify secondary phenotype variable to subset from pheno.var pheno.order}

\item{pheno.var.bg.target}{Character; specify pheno.var.2 sub-group(s) to be visualized}

\item{pheno.var.bg.colors}{Character; Colors to highlight pheno.var sub-groups ordered according to pheno.order}

\item{ax.labels}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}

\item{log}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}

\item{max.val}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}

\item{min.val}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}

\item{title}{Logical or Character; FALSE if no title, otherwise specify title in quotes}

\item{size}{Numeric; point size, default is 6}

\item{bg.size.mult}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}

\item{seed}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}
}
\value{
Scatter plot of dimensionality reduction results

Scatter plot of dimensionality reduction results
}
\description{
This function plots dimensionality reduction results and highlights cells based on known phenotype assignments

This function plots dimensionality reduction results and highlights cells based on known phenotype assignments
}
