% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_dimred.R
\name{sc_dimred}
\alias{sc_dimred}
\title{Function for plotting dimensionality reduction results and highlighting cells based on known phenotype assignments}
\usage{
sc_dimred(
  dataset = dataset,
  dimred = dimred,
  dimred.type = dimred.type,
  pheno.var = pheno.var,
  pheno.order = FALSE,
  pheno.colors = pheno.colors,
  pheno.var.2 = FALSE,
  pheno.var.2.target = pheno.var.2.target,
  ax.labels = TRUE,
  title = FALSE,
  size = 6
)
}
\arguments{
\item{dataset}{ExpressionSet object}

\item{dimred}{2-D output of t-SNE, UMAP, or PCA: Matrix with 2 columns corrersponding to dimensions 1 (x-axis) & 2 (y-axis), row numbers equivalent to the number of cells, and rownames corresponding to cell IDs}

\item{dimred.type}{Options include "tsne", "umap", or "pca"}

\item{pheno.var}{Character; Cells are highlighted according to specified phenotype variable}

\item{pheno.order}{Logical or Character; FALSE if default order; Vector of pheno.var sub-groups to be plotted}

\item{pheno.colors}{Character; Colors to highlight pheno.var sub-groups ordered according to pheno.order}

\item{pheno.var.2}{Logical or Character; FALSE if no additional subsetting, otherwise specify secondary phenotype variable to subset from pheno.var pheno.order}

\item{pheno.var.2.target}{Character; specify pheno.var.2 sub-group(s) to be visualized}

\item{ax.labels}{Logical; RUE if x- and y-axis labeled according to specified type, FALSE if no axis labels}

\item{title}{Logical or Character; FALSE if no title, otherwise specify title in quotes}

\item{size}{Numeric; point size, default is 6}
}
\value{
Scatter plot of dimensionality reduction results
}
\description{
This function plots dimensionality reduction results and highlights cells based on known phenotype assignments
}
\examples{
sc_dimred(dataset=dataset, dimred=dimred, dimred.type="umap", pheno.var=pheno.var, pheno.order=FALSE, pheno.colors=pheno.colors, pheno.var.2=FALSE, pheno.var.2.target=pheno.var.2.target, ax.labels=TRUE, title="Title", size=6)
}
